cleanup trending projects files
